+++ pkg/pyepo/model/omo/tsp.py
@@ -69,7 +69,7 @@ class tspABModel(tspABBase, optOmoModel):
         A method to solve model
         """
         self._solverfac.solve(self._model)
-        sol = np.zeros(self.num_cost, dtype=np.uint8)
+        sol = np.zeros(self.num_cost, dtype=np.float32)
         for k, (i, j) in enumerate(self.edges):
             if pe.value(self.x[i, j]) > _EDGE_ACTIVE_TOL or pe.value(self.x[j, i]) > _EDGE_ACTIVE_TOL:
                 sol[k] = 1
@@ -175,7 +175,7 @@ class tspGGModelRel(tspGGModel):
         A method to solve model — returns fractional solution.
         """
         self._solverfac.solve(self._model)
-        sol = np.zeros(self.num_cost)
+        sol = np.zeros(self.num_cost, dtype=np.float32)
         for k, (i, j) in enumerate(self.edges):
             sol[k] = pe.value(self.x[i, j]) + pe.value(self.x[j, i])
         return sol, pe.value(self._model.obj)
@@ -275,7 +275,7 @@ class tspMTZModelRel(tspMTZModel):
         A method to solve model — returns fractional solution.
         """
         self._solverfac.solve(self._model)
-        sol = np.zeros(self.num_cost)
+        sol = np.zeros(self.num_cost, dtype=np.float32)
         for k, (i, j) in enumerate(self.edges):
             sol[k] = pe.value(self.x[i, j]) + pe.value(self.x[j, i])
         return sol, pe.value(self._model.obj)
